% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_pilot_week}
\alias{add_pilot_week}
\title{Identify Pilot Week}
\usage{
add_pilot_week(df)
}
\arguments{
\item{df}{mixpanel dataframe}
}
\value{
a data frame with a new column for pilot week
}
\description{
To assist in comparisons across and within pilots, it can be useful to look
at what is happening each week. This function sets the week based on the
event's timestamp. Note that the pilot week is not calculated by the days
from the start of the pilot, but rather uses Sunday as the start of each
week.
}
\examples{
\dontrun{
df <- add_pilot_week(df)
}
}
