% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_cranberry}
\alias{dsac_cranberry}
\title{DSAC Cranberry Color}
\format{
A character string containing the hex code: "#6A1344"
}
\usage{
dsac_cranberry
}
\description{
Primary cranberry color from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
