% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_dark_navy}
\alias{dsac_dark_navy}
\title{DSAC Dark Navy Color}
\format{
A character string containing the hex code: "#123054"
}
\usage{
dsac_dark_navy
}
\description{
Dark navy variant from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
