% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_gold}
\alias{dsac_gold}
\title{DSAC Gold Color}
\format{
A character string containing the hex code: "#EFAC2F"
}
\usage{
dsac_gold
}
\description{
Accent gold color from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
