% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_light_cranberry}
\alias{dsac_light_cranberry}
\title{DSAC Light Cranberry Color}
\format{
A character string containing the hex code: "#842F66"
}
\usage{
dsac_light_cranberry
}
\description{
Light cranberry variant from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
