% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_light_navy}
\alias{dsac_light_navy}
\title{DSAC Light Navy Color}
\format{
A character string containing the hex code: "#63789D"
}
\usage{
dsac_light_navy
}
\description{
Light navy variant from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
