% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_light_teal}
\alias{dsac_light_teal}
\title{DSAC Light Teal Color}
\format{
A character string containing the hex code: "#5A9088"
}
\usage{
dsac_light_teal
}
\description{
Light teal variant from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
