% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_navy}
\alias{dsac_navy}
\title{DSAC Navy Color}
\format{
A character string containing the hex code: "#103D68"
}
\usage{
dsac_navy
}
\description{
Primary navy blue color from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
