% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_pale_navy}
\alias{dsac_pale_navy}
\title{DSAC Pale Navy Color}
\format{
A character string containing the hex code: "#C1C9D7"
}
\usage{
dsac_pale_navy
}
\description{
Pale navy variant from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
