% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_pale_teal}
\alias{dsac_pale_teal}
\title{DSAC Pale Teal Color}
\format{
A character string containing the hex code: "#D9E8E5"
}
\usage{
dsac_pale_teal
}
\description{
Pale teal variant from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
