% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsac_teal}
\alias{dsac_teal}
\title{DSAC Teal Color}
\format{
A character string containing the hex code: "#136A5D"
}
\usage{
dsac_teal
}
\description{
Primary teal color from the DSAC color palette.
}
\seealso{
\code{\link{dsac_colors}} for the complete color palette
}
\keyword{datasets}
