% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmytics-package.R
\docType{package}
\name{emmytics-package}
\alias{emmytics}
\alias{emmytics-package}
\title{emmytics: Web Analytics Tools for Eligibility Made Easy (EMMY)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides utilities for fetching, cleaning, and analyzing web analytics data from platforms like Mixpanel, with potential support for Adobe Analytics and Google Analytics. Designed for the Eligibility Made Easy (EMMY) project.
}
\author{
\strong{Maintainer}: Aaron Chafetz \email{aaron.chafetz@cms.hhs.gov} (\href{https://orcid.org/0009-0009-6460-5571}{ORCID})

}
\keyword{internal}
