% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event_properties}
\alias{event_properties}
\title{Mixpanel Events and Properties}
\format{
a data frame with 1,894 rows and 4 columns
}
\usage{
event_properties
}
\description{
This dataset provides the full set of events and the specific properties that are are
captured in Mixpanel.
}
\details{
The event are particularly useful to know when running an
API call. It is also useful to know which properties are associated with the
event, eg "ApplicantClickedCBVInvitationLink" has "seconds_since_invitation"
property. The property information can be extracted during the read in
using \code{\link{read_mixpanel}} or directly using
\code{\link{extract_properties}}.
}
\seealso{
\code{\link{read_mixpanel}}, \code{\link{extract_properties}}
}
\concept{api

\describe{
\item{event}{Mixpanel captured event}
\item{event_standard}{event name that removes providers names}
\item{key_event}{is this an event used in the user joureny,
see \code{\link{key_events}}}
\item{property}{properties stored for each event}
}}
\keyword{datasets}
