% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_applicant.R
\name{follow_applicant}
\alias{follow_applicant}
\title{Follow applicant through their session}
\usage{
follow_applicant(df, applicant, pilot_pd)
}
\arguments{
\item{df}{Mixpanel dataframe}

\item{applicant}{an applicant's distinct_id}

\item{pilot_pd}{can limit this to a particular pilot, eg "Nov 2025"}
}
\value{
a gt object
}
\description{
This function prints out a visual table of all event during an applicant's
session(s).
}
\examples{
\dontrun{
df_mp <- read_parquet(mp_path)

df_mp \%>\%
  follow_applicant("applicant-123456")
}
}
\seealso{
\code{\link{munge_journey}} for how data are munged for
\code{\link{plot_journey}} which visualizing this data

Other app_journey: 
\code{\link{munge_journey}()},
\code{\link{plot_journey}()}
}
\concept{app_journey}
