% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mixpanel_data.R
\name{get_mixpanel_data}
\alias{get_mixpanel_data}
\title{Get Mixpanel Data with Caching}
\usage{
get_mixpanel_data(
  from_date,
  to_date,
  client_agency,
  cache_dir = ".",
  force_reload = FALSE
)
}
\arguments{
\item{from_date}{Start date in 'YYYY-MM-DD' format}

\item{to_date}{End date in 'YYYY-MM-DD' format}

\item{client_agency}{if provided, will filter down by specific agency. Run
\code{unique(pilot_pds$client_agency)} to get the set of states/agencies.}

\item{cache_dir}{defaults to working directory}

\item{force_reload}{Logical. If TRUE, bypass cache and fetch fresh data}
}
\value{
A tibble containing cleaned, deduplicated Mixpanel events
}
\description{
Main function to fetch, parse, deduplicate, and cache Mixpanel data.
}
\examples{
\dontrun{
df <- get_mixpanel_data("2025-11-16", "2025-12-19")
}
}
\seealso{
Other api: 
\code{\link{cache_mixpanel}()},
\code{\link{deduplicate_mixpanel}()},
\code{\link{fetch_mixpanel}()},
\code{\link{load_cached_mixpanel}()},
\code{\link{parse_mixpanel}()},
\code{\link{standardize_properties}()}
}
\concept{api}
