% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pilot_timezones}
\alias{pilot_timezones}
\title{EMMY Pilot State Timezones}
\format{
a data frame with 4 rows and 2 columns
}
\usage{
pilot_timezones
}
\description{
Mixpanel records event in UTC. This dataset provies a mapping between the
states and their timzones to convert from UTC to local time. Uses IANA
timezone strings from \code{OlsonNames()}
}
\concept{pilot
\describe{
\item{pilot_state}{Pilot state}
\item{timezone}{IANA timezone strings from \code{OlsonNames()}}
}}
\keyword{datasets}
