% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_journey.R
\name{plot_journey}
\alias{plot_journey}
\title{Create a journey visualization plot}
\usage{
plot_journey(df, export_fldr)
}
\arguments{
\item{df}{a data frame munged by \code{\link{munge_journey}}}

\item{export_fldr}{if specified, this is the folder path where the file will
be exported to (default = missing, no export)}
}
\value{
A ggplot object
}
\description{
Create a journey visualization plot
}
\examples{
\dontrun{
df_mp <- read_parquet(mp_path)

df_mp \%>\%
  munge_journey("applicant-123456") \%>\%
  plot_journey(export = "../Images/")
}
}
\seealso{
\code{\link{munge_journey}} for how data are munged for viz

Other app_journey: 
\code{\link{munge_journey}()}
}
\concept{app_journey}
