% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mixpanel.R
\name{set_pilot}
\alias{set_pilot}
\title{Set Pilots}
\usage{
set_pilot(df)
}
\arguments{
\item{df}{data from json file}
}
\value{
dataframe with the pilot state and pilot name included
}
\description{
Standardizes the naming of the pilots and period for analysis.
}
