% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_events.R
\name{clean_events}
\alias{clean_events}
\title{Clean Mixpanel events}
\usage{
clean_events(df)
}
\arguments{
\item{df}{data frame from Mixpanel with even names}
}
\value{
dataframe with a new column for cleaned names
}
\description{
Takes the Mixpanel event names and cleans them up, removing duplicative
"Applicant" prefix and adding a space between each work. For example,
"ApplicantAccessedSuccessPage" becomes "Accessed Success Page". This is
useful when presenting out the data.
}
\examples{
\dontrun{
df_viz <- df_subset \%>\% clean_events()
}
}
