% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mixpanel_data.R
\name{get_mixpanel_data}
\alias{get_mixpanel_data}
\title{Get Mixpanel Data with Caching}
\usage{
get_mixpanel_data(
  from_date,
  to_date,
  client_agency,
  events,
  cache_dir,
  force_reload = FALSE
)
}
\arguments{
\item{from_date}{Start date in 'YYYY-MM-DD' format}

\item{to_date}{End date in 'YYYY-MM-DD' format}

\item{client_agency}{if provided, will filter down by specific agency. Run
\code{unique(pilot_pds$client_agency)} to get the set of states/agencies.}

\item{events}{Character vector of event names to fetch. If NULL (default),
all events are returned. E.g. \code{c("ApplicantViewedAgreement", 
  "ApplicantSharedIncomeSummary")}}

\item{cache_dir}{can specify a folder to store the file in (optional)}

\item{force_reload}{Logical. If TRUE, bypass cache and fetch fresh data}
}
\value{
A tibble containing cleaned, deduplicated Mixpanel events
}
\description{
Main function to fetch, parse, deduplicate, and cache Mixpanel data.
}
\examples{
\dontrun{

#pull all data over a specific time period
df_all <- get_mixpanel_data("2026-02-14", "2026-03-22")

#return just one agency's data over that period
df_la <- get_mixpanel_data("2026-02-14", "2026-03-22", 
                            client_agency = "la_ldh")
                            
#use stored information from pilot_pds
info <- pilot_pds \%>\%
  filter(state == "LA", pilot == "Feb 2026") \%>\% 
  select(start_date, end_date, client_agency) \%>\% 
  mutate(across(c(start_date, end_date), 
                \(x) as.character(x))) \%>\% 
    as.list()
  
df_la <- get_mixpanel_data(info$start_date, info$end_date,
                           client_agency = info$client_agency)
                           
#return only specific events
bounding_events <- c("ApplicantViewedAgreement", 
                     "ApplicantSharedIncomeSummary")
                     
df_la <- get_mixpanel_data(info$start_date, info$end_date,
                           client_agency = info$client_agency,
                           events = bounding_events 
                           )

}
}
\seealso{
Other api: 
\code{\link{cache_mixpanel}()},
\code{\link{deduplicate_mixpanel}()},
\code{\link{fetch_mixpanel}()},
\code{\link{load_cached_mixpanel}()},
\code{\link{parse_mixpanel}()},
\code{\link{standardize_properties}()}
}
\concept{api}
