% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_env.R
\name{load_env}
\alias{load_env}
\title{Load Environment Variables}
\usage{
load_env(
  env_path,
  req_vars = c("MIXPANEL_SERVICE_ACCOUNT_USERNAME", "MIXPANEL_SERVICE_ACCOUNT_SECRET",
    "MIXPANEL_PROJECT_ID")
)
}
\arguments{
\item{env_path}{path to .env file}

\item{req_vars}{required variables, default = c("MIXPANEL_SERVICE_ACCOUNT_USERNAME", "MIXPANEL_SERVICE_ACCOUNT_SECRET", "MIXPANEL_PROJECT_ID")}
}
\value{
named list of credentials
}
\description{
Helper to load credentials from .env file.
}
\details{
Credentials should be stored in a .env.local file in the project's root
folder. This function will use dotenv to load the variables from the
.env.local file in as environment variables to be accessed during the API
call.
}
