% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_journey.R
\name{munge_journey}
\alias{munge_journey}
\title{Munge data for journey visualization plot}
\usage{
munge_journey(df, applicant, pilot_pd)
}
\arguments{
\item{df}{Mixpanel dataframe}

\item{applicant}{an applicant's distinct_id}

\item{pilot_pd}{can limit this to a particular pilot, eg "Nov 2025"}
}
\value{
processed dataframe for use with \code{\link{plot_journey}}
}
\description{
This function provides the initial munging steps to process the data in a way
that can then be handled by \code{\link{plot_journey}}
}
\examples{
\dontrun{
df_mp <- read_parquet(mp_path)

df_mp \%>\%
  munge_journey("applicant-123456") \%>\% 
  plot_journey(export = "../Images/")
}
}
\seealso{
\code{\link{plot_journey}} for how data are munged for viz

Other app_journey: 
\code{\link{plot_journey}()}
}
\concept{app_journey}
