% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pilot_pds}
\alias{pilot_pds}
\title{EMMY State Pilot Periods}
\format{
\subsection{\code{pilot_pds}}{
}
}
\usage{
pilot_pds
}
\description{
The following dataset notes the pilot periods run in various states.
}
\seealso{
\code{\link{set_pilot}}, \code{\link{return_latest_pilot}}

A data frame with 5 rows and 5 columns
\describe{
\item{name}{Name of the pilot}
\item{state}{Two letter state abbreviation of the pilot state}
\item{client_agency}{The state agency code used in Mixpanel}
\item{start_date, end_date}{start and end period of the pilot, dates}
\item{pilot}{month and year of pilot}
}

Other pilot: 
\code{\link{return_latest_pilot}()},
\code{\link{set_pilot}()}
}
\concept{pilot}
\keyword{datasets}
