% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mixpanel.R
\name{read_mixpanel}
\alias{read_mixpanel}
\title{Read in Mixpanel data}
\usage{
read_mixpanel(file, ..., applicant_only = TRUE, drop_prop = FALSE)
}
\arguments{
\item{file}{filepath to ndjson, downloaded via \code{get_mixpanel_data}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Additional column specs
to extract from the \code{properties} column. These should be provided as
name-value pairs where the name is the new column name and the value is
the expression to extract (e.g., \code{user_id = properties$user_id}).}

\item{applicant_only}{keep only applicant data, dropping case worker and NA
default = TRUE}

\item{drop_prop}{drop the nested properties column? default = FALSE}
}
\value{
converts json to a formatted tibble
}
\description{
This function importants and applies standardized munging a Mixpanel dataset.
It expects an NDJSON file input downloaded via \code{get_mixpanel_data}. The function
extracts key fields from properties, distinct_id and cbv_flow_id, and allows
the user to provide their own fields they want to extract. It also identifies
the specific pilot periods for use in analysis. Default UTC times are
converted to local time zones.
}
\examples{
\dontrun{
# retrieve pilot data (LA Nov 2025)
get_mixpanel_data("2025-11-16", "2025-12-19", "la_ldh")

#path to json file downloaded
path <- "mixpanel_data_la_ldh_2025-11-16_to_2025-12-19.json"

#read in data with specific properties
df_la_nov <- read_mixpanel(path)
  
#read in data with specific properties
df_la_nov_extra <- read_mixpanel(path, 
        device_type = properties$device_type,
        origin = properties$origin)
  
}
}
