% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_latest_pilot}
\alias{return_latest_pilot}
\title{Return Latest Pilot}
\usage{
return_latest_pilot(df)
}
\arguments{
\item{df}{a mixpanel dataframe}
}
\value{
a character vector of the most recent pilot
}
\description{
This function returns the latest pilot in the dataset, relying on the fact
that pilot is a factor variable. Useful when there are multiple pilot periods
combined in the same dataframe.
}
\examples{
\dontrun{
#identify paths for json data for each of the periods
mp_paths <- list.files("Data","json", full.names = TRUE)

#read in data
df_mp <- mp_path |>
  set_names() |>
  map(~ read_mixpanel(.x, drop_prop = FALSE)) |>
  list_rbind(names_to = "source_path")

#store the latest pilot
latest <- return_latest_pilot(df_mp)
}
}
\seealso{
\code{\link{pilot_pds}}, \code{\link{set_pilot}}

Other pilot: 
\code{\link{pilot_pds}},
\code{\link{set_pilot}()}
}
\concept{pilot}
