% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_mixpanel.R
\name{standardize_properties}
\alias{standardize_properties}
\title{Helper function to standardize any field that might be inconsistent}
\usage{
standardize_properties(prop)
}
\arguments{
\item{prop}{property/field to standarized}
}
\description{
Helper function to standardize any field that might be inconsistent
}
\keyword{internal}
